# docs/openapi/SCIMServer_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SCIMServer API",
    "version": "0.37.3",
    "description": "Complete SCIM 2.0 Server - 82 endpoints across 19 controllers.\n\nMulti-tenant endpoint architecture, OAuth 2.0, per-endpoint credentials, bulk operations, /Me endpoint, custom resource types, database browser, activity feed, per-endpoint logs, structured logging.\n\nAll SCIM paths are scoped under /endpoints/{endpointId}. Admin paths are under /admin.\nGlobal prefix: /scim (configurable via API_PREFIX). URL rewriting: /scim/v2/* → /scim/*.",
    "contact": { "name": "SCIMServer", "url": "https://github.com/pranems/SCIMServer" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000/scim", "description": "Local development (default)" },
    { "url": "http://localhost:6000/scim", "description": "Local (port 6000)" },
    { "url": "http://localhost:8080/scim", "description": "Docker" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Shared secret, per-endpoint credential, or OAuth JWT" },
      "oauthJwt": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "OAuth 2.0 JWT with sub claim - required for /Me" }
    },
    "schemas": {
      "ScimMeta": {
        "type": "object",
        "properties": {
          "resourceType": { "type": "string", "example": "User" },
          "created": { "type": "string", "format": "date-time" },
          "lastModified": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "version": { "type": "string", "example": "W/\"1\"" }
        }
      },
      "ScimUser": {
        "type": "object",
        "required": ["schemas", "userName"],
        "properties": {
          "schemas": { "type": "array", "items": { "type": "string" } },
          "id": { "type": "string", "format": "uuid", "readOnly": true },
          "userName": { "type": "string" },
          "externalId": { "type": "string" },
          "active": { "type": "boolean" },
          "name": { "type": "object", "properties": { "givenName": { "type": "string" }, "familyName": { "type": "string" }, "formatted": { "type": "string" } } },
          "displayName": { "type": "string" },
          "nickName": { "type": "string" },
          "title": { "type": "string" },
          "userType": { "type": "string" },
          "preferredLanguage": { "type": "string" },
          "locale": { "type": "string" },
          "timezone": { "type": "string" },
          "emails": { "type": "array", "items": { "type": "object", "properties": { "value": { "type": "string" }, "type": { "type": "string" }, "primary": { "type": "boolean" } } } },
          "phoneNumbers": { "type": "array", "items": { "type": "object", "properties": { "value": { "type": "string" }, "type": { "type": "string" } } } },
          "addresses": { "type": "array", "items": { "type": "object", "properties": { "streetAddress": { "type": "string" }, "locality": { "type": "string" }, "region": { "type": "string" }, "postalCode": { "type": "string" }, "country": { "type": "string" }, "type": { "type": "string" }, "primary": { "type": "boolean" } } } },
          "groups": { "type": "array", "readOnly": true, "items": { "type": "object", "properties": { "value": { "type": "string" }, "display": { "type": "string" } } } },
          "meta": { "$ref": "#/components/schemas/ScimMeta" }
        },
        "additionalProperties": true
      },
...
      "PatchOp": {
        "type": "object",
        "required": ["schemas", "Operations"],
        "properties": {
          "schemas": { "type": "array", "items": { "type": "string" } },
          "Operations": { "type": "array", "maxItems": 1000, "items": { "type": "object", "required": ["op"], "properties": { "op": { "type": "string", "enum": ["add", "replace", "remove"] }, "path": { "type": "string" }, "value": {} } } }
        }
      },
...
      "Endpoint": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "displayName": { "type": "string" },
          "description": { "type": "string" },
          "active": { "type": "boolean" },
          "profile": { "type": "object", "description": "Full profile (view=full)" },
          "profileSummary": { "type": "object", "description": "Profile summary (view=summary)" },
          "scimBasePath": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "_links": { "type": "object", "properties": { "self": { "type": "string" }, "stats": { "type": "string" }, "credentials": { "type": "string" }, "scim": { "type": "string" } } }
        }
      },
...
      "UpdateEndpoint": { "type": "object", "properties": { "displayName": { "type": "string" }, "description": { "type": "string" }, "active": { "type": "boolean" }, "profile": { "type": "object" } } },
...
    },
    "parameters": {
      "endpointId": { "name": "endpointId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Endpoint (tenant) UUID" },
...
      "attributes": { "name": "attributes", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated attributes to include" },
      "excludedAttributes": { "name": "excludedAttributes", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated attributes to exclude" },
...
      "view": { "name": "view", "in": "query", "schema": { "type": "string", "enum": ["summary", "full"] } }
    }
  },
  "paths": {
...
    "/endpoints/{endpointId}/Me": {
      "get": { "tags": ["Me"], "summary": "Get current user", "operationId": "getMe", "description": "Requires OAuth JWT with sub claim matching a userName", "security": [{ "oauthJwt": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "$ref": "#/components/parameters/attributes" }, { "$ref": "#/components/parameters/excludedAttributes" }], "responses": { "200": { "description": "Current user" }, "401": { "description": "Not OAuth" }, "404": { "description": "No matching user" } } },
      "put": { "tags": ["Me"], "summary": "Replace current user", "operationId": "replaceMe", "security": [{ "oauthJwt": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "$ref": "#/components/parameters/attributes" }, { "$ref": "#/components/parameters/excludedAttributes" }], "requestBody": { "required": true, "content": { "application/scim+json": { "schema": { "$ref": "#/components/schemas/ScimUser" } } } }, "responses": { "200": { "description": "Replaced" }, "401": { "description": "Not OAuth" }, "404": { "description": "Not found" } } },
      "patch": { "tags": ["Me"], "summary": "Patch current user", "operationId": "patchMe", "security": [{ "oauthJwt": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "$ref": "#/components/parameters/attributes" }, { "$ref": "#/components/parameters/excludedAttributes" }], "requestBody": { "required": true, "content": { "application/scim+json": { "schema": { "$ref": "#/components/schemas/PatchOp" } } } }, "responses": { "200": { "description": "Patched" }, "401": { "description": "Not OAuth" } } },
      "delete": { "tags": ["Me"], "summary": "Delete current user", "operationId": "deleteMe", "security": [{ "oauthJwt": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "responses": { "204": { "description": "Deleted" }, "401": { "description": "Not OAuth" } } }
    },
...
    "/admin/endpoints/{endpointId}": {
      "get": { "tags": ["Admin - Endpoints"], "summary": "Get endpoint by ID", "operationId": "getEndpoint", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "$ref": "#/components/parameters/view" }], "responses": { "200": { "description": "Endpoint", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Endpoint" } } } }, "404": { "description": "Not found" } } },
      "patch": { "tags": ["Admin - Endpoints"], "summary": "Update endpoint (deep-merges settings)", "operationId": "updateEndpoint", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEndpoint" } } } }, "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } } },
      "delete": { "tags": ["Admin - Endpoints"], "summary": "Delete endpoint (cascades all data)", "operationId": "deleteEndpoint", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
...
  }
}